requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
opencv-python==4.7.0.72
matplotlib==3.7.1
seaborn==0.12.2
streamlit==1.22.0
Pillow==9.5.0 